Clean up CMakeLists
The VERSION and SOVERSION were set to the same, making the soname of the
library contain the minor and patch numbers. This is now fixed, so that
the soname only contains the major version. This also fixed the
libpredict.so.<major> symlink issues, so manually symlinking isn't
necessary anymore.

Remove the pkg-config file from the lib directory. The file is
architecture independent and should live in the share directory.