Fix license and credits
Update the COPYING file to the latest version of GPL version 2 to get
the correct address of the FSF. This file was downloaded from the FSF
website.

Rewrite CREDITS to refer to predict, and delete all the stuff about
other parts of predict, most of which isn't part of libpredict.

Update README.md with copyright and licensing information. This
licensing clause is expected to cover the whole source tree.